Extracting the instance variable setting to a separate module and including it in the local_errors_controller.
